rename settings store
